Implement email on bullmq queue
